pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = "Scripts for DMTA"
authors = ["darkman <darkman@meganet.ru>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
colorlog = "^6.4.1"
requests = "^2.26.0"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
